requirements.txt
# ==================== CORE SYSTEM PACKAGES (install first) ====================
typing-extensions==4.9.0
urllib3==2.1.0

# ==================== CRYPTOGRAPHY & SECURITY ====================
cryptography==42.0.0

# ==================== TIMEZONE & DATE HANDLING ====================
pytz==2024.1

# ==================== ENVIRONMENT VARIABLES ====================
python-dotenv==1.0.0

# ==================== NUMERICAL & DATA PROCESSING ====================
numpy==1.26.3
pandas==2.1.4

# ==================== HTTP & NETWORKING ====================
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.3

# ==================== DATABASE ====================
supabase==2.3.0

# ==================== SCHEDULING ====================
APScheduler==3.10.4

# ==================== MACHINE LEARNING ====================
scikit-learn==1.4.0
xgboost==2.0.3

# ==================== DEEP LEARNING (LSTM) ====================
tensorflow==2.15.0

# ==================== VISUALIZATION ====================
matplotlib==3.8.2
mplfinance==0.12.10b0

# ==================== LOGGING ====================
colorlog==6.8.2

# ==================== TELEGRAM BOT (install last - has many dependencies) ====================
python-telegram-bot==20.7

# ==================== MT5 WORKER SERVICE (Windows VPS only) ====================
# DO NOT install these on your Linux bot server
# Only install on Windows VPS where mt5_worker.py runs:
#   pip install MetaTrader5 flask flask-limiter pywin32

aiohttp
waitress
